+++ packages/ai/src/generate-text/step-result.ts
@@ -1,3 +1,8 @@
+import type {
+  Context,
+  InferToolSetContext,
+  ToolSet,
+} from '@ai-sdk/provider-utils';
 import { ReasoningFilePart, ReasoningPart } from '@ai-sdk/provider-utils';
 import {
   CallWarning,
@@ -23,15 +28,13 @@ import {
   StaticToolResult,
   TypedToolResult,
 } from './tool-result';
-import type { GenerationContext } from './generation-context';
-import type { ToolSet } from '@ai-sdk/provider-utils';
 
 /**
  * The result of a single step in the generation process.
  */
 export type StepResult<
   TOOLS extends ToolSet,
-  CONTEXT extends GenerationContext<TOOLS>,
+  USER_CONTEXT extends Context = Context,
 > = {
   /**
    * Unique identifier for the generation call this step belongs to.
@@ -68,7 +71,7 @@ export type StepResult<
    *
    * Experimental (can break in patch releases).
    */
-  readonly context: CONTEXT;
+  readonly context: InferToolSetContext<TOOLS> & USER_CONTEXT;
 
   /**
    * The content that was generated in the last step.
@@ -182,22 +185,25 @@ export type StepResult<
 
 export class DefaultStepResult<
   TOOLS extends ToolSet,
-  CONTEXT extends GenerationContext<TOOLS>,
-> implements StepResult<TOOLS, CONTEXT> {
-  readonly callId: StepResult<TOOLS, CONTEXT>['callId'];
-  readonly stepNumber: StepResult<TOOLS, CONTEXT>['stepNumber'];
-  readonly model: StepResult<TOOLS, CONTEXT>['model'];
-  readonly functionId: StepResult<TOOLS, CONTEXT>['functionId'];
-  readonly metadata: StepResult<TOOLS, CONTEXT>['metadata'];
-  readonly context: StepResult<TOOLS, CONTEXT>['context'];
-  readonly content: StepResult<TOOLS, CONTEXT>['content'];
-  readonly finishReason: StepResult<TOOLS, CONTEXT>['finishReason'];
-  readonly rawFinishReason: StepResult<TOOLS, CONTEXT>['rawFinishReason'];
-  readonly usage: StepResult<TOOLS, CONTEXT>['usage'];
-  readonly warnings: StepResult<TOOLS, CONTEXT>['warnings'];
-  readonly request: StepResult<TOOLS, CONTEXT>['request'];
-  readonly response: StepResult<TOOLS, CONTEXT>['response'];
-  readonly providerMetadata: StepResult<TOOLS, CONTEXT>['providerMetadata'];
+  USER_CONTEXT extends Context = Context,
+> implements StepResult<TOOLS, USER_CONTEXT> {
+  readonly callId: StepResult<TOOLS, USER_CONTEXT>['callId'];
+  readonly stepNumber: StepResult<TOOLS, USER_CONTEXT>['stepNumber'];
+  readonly model: StepResult<TOOLS, USER_CONTEXT>['model'];
+  readonly functionId: StepResult<TOOLS, USER_CONTEXT>['functionId'];
+  readonly metadata: StepResult<TOOLS, USER_CONTEXT>['metadata'];
+  readonly context: StepResult<TOOLS, USER_CONTEXT>['context'];
+  readonly content: StepResult<TOOLS, USER_CONTEXT>['content'];
+  readonly finishReason: StepResult<TOOLS, USER_CONTEXT>['finishReason'];
+  readonly rawFinishReason: StepResult<TOOLS, USER_CONTEXT>['rawFinishReason'];
+  readonly usage: StepResult<TOOLS, USER_CONTEXT>['usage'];
+  readonly warnings: StepResult<TOOLS, USER_CONTEXT>['warnings'];
+  readonly request: StepResult<TOOLS, USER_CONTEXT>['request'];
+  readonly response: StepResult<TOOLS, USER_CONTEXT>['response'];
+  readonly providerMetadata: StepResult<
+    TOOLS,
+    USER_CONTEXT
+  >['providerMetadata'];
 
   constructor({
     callId,
@@ -216,21 +222,21 @@ export class DefaultStepResult<
     response,
     providerMetadata,
   }: {
-    callId: StepResult<TOOLS, CONTEXT>['callId'];
-    stepNumber: StepResult<TOOLS, CONTEXT>['stepNumber'];
-    provider: StepResult<TOOLS, CONTEXT>['model']['provider'];
-    modelId: StepResult<TOOLS, CONTEXT>['model']['modelId'];
-    functionId: StepResult<TOOLS, CONTEXT>['functionId'];
-    metadata: StepResult<TOOLS, CONTEXT>['metadata'];
-    context: StepResult<TOOLS, CONTEXT>['context'];
-    content: StepResult<TOOLS, CONTEXT>['content'];
-    finishReason: StepResult<TOOLS, CONTEXT>['finishReason'];
-    rawFinishReason: StepResult<TOOLS, CONTEXT>['rawFinishReason'];
-    usage: StepResult<TOOLS, CONTEXT>['usage'];
-    warnings: StepResult<TOOLS, CONTEXT>['warnings'];
-    request: StepResult<TOOLS, CONTEXT>['request'];
-    response: StepResult<TOOLS, CONTEXT>['response'];
-    providerMetadata: StepResult<TOOLS, CONTEXT>['providerMetadata'];
+    callId: StepResult<TOOLS, USER_CONTEXT>['callId'];
+    stepNumber: StepResult<TOOLS, USER_CONTEXT>['stepNumber'];
+    provider: StepResult<TOOLS, USER_CONTEXT>['model']['provider'];
+    modelId: StepResult<TOOLS, USER_CONTEXT>['model']['modelId'];
+    functionId: StepResult<TOOLS, USER_CONTEXT>['functionId'];
+    metadata: StepResult<TOOLS, USER_CONTEXT>['metadata'];
+    context: StepResult<TOOLS, USER_CONTEXT>['context'];
+    content: StepResult<TOOLS, USER_CONTEXT>['content'];
+    finishReason: StepResult<TOOLS, USER_CONTEXT>['finishReason'];
+    rawFinishReason: StepResult<TOOLS, USER_CONTEXT>['rawFinishReason'];
+    usage: StepResult<TOOLS, USER_CONTEXT>['usage'];
+    warnings: StepResult<TOOLS, USER_CONTEXT>['warnings'];
+    request: StepResult<TOOLS, USER_CONTEXT>['request'];
+    response: StepResult<TOOLS, USER_CONTEXT>['response'];
+    providerMetadata: StepResult<TOOLS, USER_CONTEXT>['providerMetadata'];
   }) {
     this.callId = callId;
     this.stepNumber = stepNumber;