+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/anthropic
 
+## 4.0.0-beta.25
+
+### Patch Changes
+
+- 09bd27b: feat (provider/anthropic): add support for inference_geo provider option
+
+## 4.0.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [3ae1786]
+  - @ai-sdk/provider-utils@5.0.0-beta.17
+
 ## 4.0.0-beta.23
 
 ### Patch Changes